refactors: move cli app to seperate module
